requirements.txt
asgiref==3.3.1
astroid==2.5.2
bandit==1.7.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
colorama==0.4.4
coverage==5.5
Django==2.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
Flask==1.1.2
gitdb==4.0.7
GitPython==3.1.14
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pbr==5.5.1
Pillow==8.1.2
pycparser==2.20
pylint==2.7.4
pytz==2021.1
PyYAML==5.4.1
rope==0.18.0
six==1.15.0
smmap==4.0.0
sqlparse==0.4.1
stevedore==3.3.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
